improvements to scripts, loading of bad waveform trimmings, and plotting
